logs

+++ proj-1/xmod.c
@@ -24,7 +24,7 @@ __mode_t get_perms(unsigned int r, unsigned int w, unsigned int x, char op, char
 void chmod_dir(char* cmd, char* dir_name, int verbosity, int argc, char* argv[], char* envp[]);
 char * formatOctal(char *octal);
 void strmode(__mode_t mode, char * buf);
-double get_running_time();
+long int get_running_time();
 void set_log_file(FILE* logFile);
 void log_start();
